Create webApp.yaml
This web app consists of an app within a single region, a VPC with an IGW, two availability zones, a public subnet in each, an app subnet with an EC2 instance in each, and a DB subnet with an RDS instance in each. This first commit includes the template, and build of the VPC resource to start.